+++ src/Application/MediatR/PlannedMeal/Queries/DoesPlannedMealExistByDateAndOrdinalNumberQuery.cs
@@ -7,5 +7,5 @@ using System.Threading.Tasks;
 
 namespace FoodPlanner.Application.MediatR.PlannedMeal.Queries
 {
-    public record DoesPlannedMealExistByDateAndOrdinalNumberQuery(DateTime ScheduledFor, uint OridinalNumber) : IRequest<bool>;
+    public record DoesPlannedMealExistByDateAndOrdinalNumberQuery(DateTime ScheduledFor, byte OridinalNumber) : IRequest<bool>;
 }
